requirements.txt
elasticsearch==8.15.0

uvicorn==0.25.0
fastapi==0.109.0
requests==2.31.0
itsdangerous
load_dotenv==0.1.0
numpy==1.26.0
pandas==1.1.5
nltk==3.9.1

## LLM Requirements
openai==1.42.0
httpx==0.26.0
google-generativeai   
#aiohttp==3.9.3

## UI Requirements
gradio==3.50.2